requirements.txt
openai==0.28
discord.py
python-dotenv
langchain
langchain-community
faiss-cpu
sentence-transformers
numpy